www/django_portfolio/portfolio_website/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
cmake==3.27.7
Django==4.2.5
einops==0.7.0
filelock==3.12.3
fsspec==2023.10.0
idna==3.4
Jinja2==3.1.2
lit==17.0.5
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
Pillow==10.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
sympy==1.12
torch==2.1.1
torchvision==0.16.1
triton==2.1.0
typing_extensions==4.7.1
urllib3==2.1.0
